Convert wait-for-advert to TypeScript
