--- static/src/javascripts/projects/commercial/modules/dfp/wait-for-advert.js
@@ -1,17 +1,0 @@
-import { memoize } from 'lodash-es';
-import { getAdvertById } from './get-advert-by-id';
-
-export const waitForAdvert = memoize(
-    (id) =>
-        new Promise(resolve => {
-            const checkAdvert = () => {
-                const advert = getAdvertById(id);
-                if (!advert) {
-                    window.setTimeout(checkAdvert, 200);
-                } else {
-                    resolve(advert);
-                }
-            };
-            checkAdvert();
-        })
-);
